animations
git-svn-id: svn+ssh://fileserverness.ath.cx/serv/svn/virus-td@201 05766cc9-4f33-4ba7-801d-bd015708efd9

+++ Tower.h
@@ -31,7 +31,7 @@ protected:
   Particles * weapon;// Particle
   char* sound;
   ProjectileManager* projectiles;
-  int animateSpeed; // animation speed
+  double animateSpeed; // animation speed
   //shadow * the_shadow;
 public:
   Tower(float inx, float iny, float inz, int gx, int gy);
